pyproject.toml
[project]
name = "certificate-service-api"
version = "0.1.0"
description = ""
authors = [
    {name = "DoeunKongden",email = "doeunkongden1@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13, <4"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn[standard] (>=0.40.0,<0.41.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "greenlet (>=3.3.1,<4.0.0)",
    "redis (>=7.1.0,<8.0.0)",
    "pydantic[email] (>=2.12.5,<3.0.0)",
    "python-keycloak (>=7.0.3,<8.0.0)"
]

[tool.poetry]
packages = [{include = "app"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
